stable version of Solomka

+++ runtime/src/cache_hash_data.rs
@@ -1,8 +1,10 @@
 //! Cached data for hashing accounts
-#[cfg(test)]
-use crate::pubkey_bins::PubkeyBinCalculator24;
 use {
-    crate::{accounts_hash::CalculateHashIntermediate, cache_hash_data_stats::CacheHashDataStats},
+    crate::{
+        accounts_hash::CalculateHashIntermediate, cache_hash_data_stats::CacheHashDataStats,
+        pubkey_bins::PubkeyBinCalculator24,
+    },
+    log::*,
     memmap2::MmapMut,
     solana_measure::measure::Measure,
     std::{
@@ -23,74 +25,16 @@ pub struct Header {
     count: usize,
 }
 
-pub(crate) struct CacheHashDataFile {
+struct CacheHashDataFile {
     cell_size: u64,
     mmap: MmapMut,
     capacity: u64,
 }
 
 impl CacheHashDataFile {
-    /// return a slice of a reference to all the cache hash data from the mmapped file
-    pub(crate) fn get_cache_hash_data(&self) -> &[EntryType] {
-        self.get_slice(0)
-    }
-
-    #[cfg(test)]
-    /// Populate 'accumulator' from entire contents of the cache file.
-    pub(crate) fn load_all(
-        &self,
-        accumulator: &mut SavedType,
-        start_bin_index: usize,
-        bin_calculator: &PubkeyBinCalculator24,
-        stats: &mut CacheHashDataStats,
-    ) {
-        let mut m2 = Measure::start("decode");
-        let slices = self.get_cache_hash_data();
-        for d in slices {
-            let mut pubkey_to_bin_index = bin_calculator.bin_from_pubkey(&d.pubkey);
-            assert!(
-                pubkey_to_bin_index >= start_bin_index,
-                "{pubkey_to_bin_index}, {start_bin_index}"
-            ); // this would indicate we put a pubkey in too high of a bin
-            pubkey_to_bin_index -= start_bin_index;
-            accumulator[pubkey_to_bin_index].push(d.clone()); // may want to avoid clone here
-        }
-
-        m2.stop();
-        stats.decode_us += m2.as_us();
-    }
-
-    /// get '&mut EntryType' from cache file [ix]
-    fn get_mut(&mut self, ix: u64) -> &mut EntryType {
-        let item_slice = self.get_slice_internal(ix);
-        unsafe {
-            let item = item_slice.as_ptr() as *mut EntryType;
-            &mut *item
-        }
-    }
-
-    /// get '&[EntryType]' from cache file [ix..]
-    fn get_slice(&self, ix: u64) -> &[EntryType] {
-        let start = self.get_element_offset_byte(ix);
-        let item_slice: &[u8] = &self.mmap[start..];
-        let remaining_elements = item_slice.len() / std::mem::size_of::<EntryType>();
-        unsafe {
-            let item = item_slice.as_ptr() as *const EntryType;
-            std::slice::from_raw_parts(item, remaining_elements)
-        }
-    }
-
-    /// return byte offset of entry 'ix' into a slice which contains a header and at least ix elements
-    fn get_element_offset_byte(&self, ix: u64) -> usize {
+    fn get_mut<T: Sized>(&mut self, ix: u64) -> &mut T {
         let start = (ix * self.cell_size) as usize + std::mem::size_of::<Header>();
-        debug_assert_eq!(start % std::mem::align_of::<EntryType>(), 0);
-        start
-    }
-
-    /// get the bytes representing cache file [ix]
-    fn get_slice_internal(&self, ix: u64) -> &[u8] {
-        let start = self.get_element_offset_byte(ix);
-        let end = start + std::mem::size_of::<EntryType>();
+        let end = start + std::mem::size_of::<T>();
         assert!(
             end <= self.capacity as usize,
             "end: {}, capacity: {}, ix: {}, cell size: {}",
@@ -99,7 +43,11 @@ impl CacheHashDataFile {
             ix,
             self.cell_size
         );
-        &self.mmap[start..end]
+        let item_slice: &[u8] = &self.mmap[start..end];
+        unsafe {
+            let item = item_slice.as_ptr() as *mut T;
+            &mut *item
+        }
     }
 
     fn get_header_mut(&mut self) -> &mut Header {
@@ -112,7 +60,7 @@ impl CacheHashDataFile {
         }
     }
 
-    fn new_map(file: impl AsRef<Path>, capacity: u64) -> Result<MmapMut, std::io::Error> {
+    fn new_map(file: &Path, capacity: u64) -> Result<MmapMut, std::io::Error> {
         let mut data = OpenOptions::new()
             .read(true)
             .write(true)
@@ -124,12 +72,12 @@ impl CacheHashDataFile {
         // expensive.
         data.seek(SeekFrom::Start(capacity - 1)).unwrap();
         data.write_all(&[0]).unwrap();
-        data.rewind().unwrap();
+        data.seek(SeekFrom::Start(0)).unwrap();
         data.flush().unwrap();
         Ok(unsafe { MmapMut::map_mut(&data).unwrap() })
     }
 
-    fn load_map(file: impl AsRef<Path>) -> Result<MmapMut, std::io::Error> {
+    fn load_map(file: &Path) -> Result<MmapMut, std::io::Error> {
         let data = OpenOptions::new()
             .read(true)
             .write(true)
@@ -140,9 +88,9 @@ impl CacheHashDataFile {
     }
 }
 
-pub type PreExistingCacheFiles = HashSet<PathBuf>;
+pub type PreExistingCacheFiles = HashSet<String>;
 pub struct CacheHashData {
-    cache_dir: PathBuf,
+    cache_folder: PathBuf,
     pre_existing_cache_files: Arc<Mutex<PreExistingCacheFiles>>,
     pub stats: Arc<Mutex<CacheHashDataStats>>,
 }
@@ -155,13 +103,14 @@ impl Drop for CacheHashData {
 }
 
 impl CacheHashData {
-    pub fn new(cache_dir: PathBuf) -> CacheHashData {
-        std::fs::create_dir_all(&cache_dir).unwrap_or_else(|err| {
-            panic!("error creating cache dir {}: {err}", cache_dir.display())
-        });
+    pub fn new<P: AsRef<Path> + std::fmt::Debug>(parent_folder: &P) -> CacheHashData {
+        let cache_folder = Self::get_cache_root_path(parent_folder);
+
+        std::fs::create_dir_all(cache_folder.clone())
+            .unwrap_or_else(|_| panic!("error creating cache dir: {:?}", cache_folder));
 
         let result = CacheHashData {
-            cache_dir,
+            cache_folder,
             pre_existing_cache_files: Arc::new(Mutex::new(PreExistingCacheFiles::default())),
             stats: Arc::new(Mutex::new(CacheHashDataStats::default())),
         };
@@ -174,19 +123,19 @@ impl CacheHashData {
         if !pre_existing_cache_files.is_empty() {
             self.stats.lock().unwrap().unused_cache_files += pre_existing_cache_files.len();
             for file_name in pre_existing_cache_files.iter() {
-                let result = self.cache_dir.join(file_name);
+                let result = self.cache_folder.join(file_name);
                 let _ = fs::remove_file(result);
             }
         }
     }
     fn get_cache_files(&self) {
-        if self.cache_dir.is_dir() {
-            let dir = fs::read_dir(&self.cache_dir);
+        if self.cache_folder.is_dir() {
+            let dir = fs::read_dir(self.cache_folder.clone());
             if let Ok(dir) = dir {
                 let mut pre_existing = self.pre_existing_cache_files.lock().unwrap();
                 for entry in dir.flatten() {
                     if let Some(name) = entry.path().file_name() {
-                        pre_existing.insert(PathBuf::from(name));
+                        pre_existing.insert(name.to_str().unwrap().to_string());
                     }
                 }
                 self.stats.lock().unwrap().cache_file_count += pre_existing.len();
@@ -194,43 +143,41 @@ impl CacheHashData {
         }
     }
 
-    #[cfg(test)]
+    fn get_cache_root_path<P: AsRef<Path>>(parent_folder: &P) -> PathBuf {
+        parent_folder.as_ref().join("calculate_accounts_hash_cache")
+    }
+
     /// load from 'file_name' into 'accumulator'
-    pub(crate) fn load(
+    pub fn load<P: AsRef<Path> + std::fmt::Debug>(
         &self,
-        file_name: impl AsRef<Path>,
+        file_name: &P,
         accumulator: &mut SavedType,
         start_bin_index: usize,
         bin_calculator: &PubkeyBinCalculator24,
     ) -> Result<(), std::io::Error> {
-        let mut m = Measure::start("overall");
-        let cache_file = self.load_map(file_name)?;
         let mut stats = CacheHashDataStats::default();
-        cache_file.load_all(accumulator, start_bin_index, bin_calculator, &mut stats);
-        m.stop();
-        self.stats.lock().unwrap().load_us += m.as_us();
-        Ok(())
-    }
-
-    /// map 'file_name' into memory
-    pub(crate) fn load_map(
-        &self,
-        file_name: impl AsRef<Path>,
-    ) -> Result<CacheHashDataFile, std::io::Error> {
-        let mut stats = CacheHashDataStats::default();
-        let result = self.map(file_name, &mut stats);
-        self.stats.lock().unwrap().accumulate(&stats);
+        let result = self.load_internal(
+            file_name,
+            accumulator,
+            start_bin_index,
+            bin_calculator,
+            &mut stats,
+        );
+        self.stats.lock().unwrap().merge(&stats);
         result
     }
 
-    /// create and return a MappedCacheFile for a cache file path
-    fn map(
+    fn load_internal<P: AsRef<Path> + std::fmt::Debug>(
         &self,
-        file_name: impl AsRef<Path>,
+        file_name: &P,
+        accumulator: &mut SavedType,
+        start_bin_index: usize,
+        bin_calculator: &PubkeyBinCalculator24,
         stats: &mut CacheHashDataStats,
-    ) -> Result<CacheHashDataFile, std::io::Error> {
-        let path = self.cache_dir.join(&file_name);
-        let file_len = std::fs::metadata(&path)?.len();
+    ) -> Result<(), std::io::Error> {
+        let mut m = Measure::start("overall");
+        let path = self.cache_folder.join(file_name);
+        let file_len = std::fs::metadata(path.clone())?.len();
         let mut m1 = Measure::start("read_file");
         let mmap = CacheHashDataFile::load_map(&path)?;
         m1.stop();
@@ -241,14 +188,6 @@ impl CacheHashData {
         }
 
         let cell_size = std::mem::size_of::<EntryType>() as u64;
-        unsafe {
-            assert_eq!(
-                mmap.align_to::<EntryType>().0.len(),
-                0,
-                "mmap is not aligned"
-            );
-        }
-        assert_eq!((cell_size as usize) % std::mem::size_of::<u64>(), 0);
         let mut cache_file = CacheHashDataFile {
             mmap,
             cell_size,
@@ -264,45 +203,70 @@ impl CacheHashData {
         cache_file.capacity = capacity;
         assert_eq!(
             capacity, file_len,
-            "expected: {capacity}, len on disk: {file_len} {}, entries: {entries}, cell_size: {cell_size}", path.display(),
+            "expected: {}, len on disk: {} {:?}, entries: {}, cell_size: {}",
+            capacity, file_len, path, entries, cell_size
         );
 
         stats.total_entries = entries;
         stats.cache_file_size += capacity as usize;
 
-        self.pre_existing_cache_files
+        let file_name_lookup = file_name.as_ref().to_str().unwrap().to_string();
+        let found = self
+            .pre_existing_cache_files
             .lock()
             .unwrap()
-            .remove(file_name.as_ref());
+            .remove(&file_name_lookup);
+        if !found {
+            info!(
+                "tried to mark {:?} as used, but it wasn't in the set, one example: {:?}",
+                file_name_lookup,
+                self.pre_existing_cache_files.lock().unwrap().iter().next()
+            );
+        }
 
         stats.loaded_from_cache += 1;
         stats.entries_loaded_from_cache += entries;
+        let mut m2 = Measure::start("decode");
+        for i in 0..entries {
+            let d = cache_file.get_mut::<EntryType>(i as u64);
+            let mut pubkey_to_bin_index = bin_calculator.bin_from_pubkey(&d.pubkey);
+            assert!(
+                pubkey_to_bin_index >= start_bin_index,
+                "{}, {}",
+                pubkey_to_bin_index,
+                start_bin_index
+            ); // this would indicate we put a pubkey in too high of a bin
+            pubkey_to_bin_index -= start_bin_index;
+            accumulator[pubkey_to_bin_index].push(d.clone()); // may want to avoid clone here
+        }
 
-        Ok(cache_file)
+        m2.stop();
+        stats.decode_us += m2.as_us();
+        m.stop();
+        stats.load_us += m.as_us();
+        Ok(())
     }
 
     /// save 'data' to 'file_name'
-    pub fn save(
-        &self,
-        file_name: impl AsRef<Path>,
-        data: &SavedTypeSlice,
-    ) -> Result<(), std::io::Error> {
+    pub fn save(&self, file_name: &Path, data: &SavedTypeSlice) -> Result<(), std::io::Error> {
         let mut stats = CacheHashDataStats::default();
         let result = self.save_internal(file_name, data, &mut stats);
-        self.stats.lock().unwrap().accumulate(&stats);
+        self.stats.lock().unwrap().merge(&stats);
         result
     }
 
     fn save_internal(
         &self,
-        file_name: impl AsRef<Path>,
+        file_name: &Path,
         data: &SavedTypeSlice,
         stats: &mut CacheHashDataStats,
     ) -> Result<(), std::io::Error> {
         let mut m = Measure::start("save");
-        let cache_path = self.cache_dir.join(file_name);
-        // overwrite any existing file at this path
-        let _ignored = remove_file(&cache_path);
+        let cache_path = self.cache_folder.join(file_name);
+        let create = true;
+        if create {
+            let _ignored = remove_file(&cache_path);
+        }
         let cell_size = std::mem::size_of::<EntryType>() as u64;
         let mut m1 = Measure::start("create save");
         let entries = data
@@ -321,7 +285,7 @@ impl CacheHashData {
             capacity,
         };
 
-        let header = cache_file.get_header_mut();
+        let mut header = cache_file.get_header_mut();
         header.count = entries;
 
         stats.cache_file_size = capacity as usize;
@@ -331,7 +295,7 @@ impl CacheHashData {
         let mut i = 0;
         data.iter().for_each(|x| {
             x.iter().for_each(|item| {
-                let d = cache_file.get_mut(i as u64);
+                let d = cache_file.get_mut::<EntryType>(i as u64);
                 i += 1;
                 *d = item.clone();
             })
@@ -358,8 +322,7 @@ pub mod tests {
         // compare
         use tempfile::TempDir;
         let tmpdir = TempDir::new().unwrap();
-        let cache_dir = tmpdir.path().to_path_buf();
-        std::fs::create_dir_all(&cache_dir).unwrap();
+        std::fs::create_dir_all(&tmpdir).unwrap();
 
         for bins in [1, 2, 4] {
             let bin_calculator = PubkeyBinCalculator24::new(bins);
@@ -386,9 +349,10 @@ pub mod tests {
                                 data_this_pass.push(this_bin_data);
                             }
                         }
-                        let cache = CacheHashData::new(cache_dir.clone());
-                        let file_name = PathBuf::from("test");
-                        cache.save(&file_name, &data_this_pass).unwrap();
+                        let cache = CacheHashData::new(&tmpdir);
+                        let file_name = "test";
+                        let file = Path::new(file_name).to_path_buf();
+                        cache.save(&file, &data_this_pass).unwrap();
                         cache.get_cache_files();
                         assert_eq!(
                             cache
@@ -397,11 +361,11 @@ pub mod tests {
                                 .unwrap()
                                 .iter()
                                 .collect::<Vec<_>>(),
-                            vec![&file_name],
+                            vec![file_name]
                         );
-                        let mut accum = (0..bins_per_pass).map(|_| vec![]).collect();
+                        let mut accum = (0..bins_per_pass).into_iter().map(|_| vec![]).collect();
                         cache
-                            .load(&file_name, &mut accum, start_bin_this_pass, &bin_calculator)
+                            .load(&file, &mut accum, start_bin_this_pass, &bin_calculator)
                             .unwrap();
                         if flatten_data {
                             bin_data(
@@ -413,7 +377,8 @@ pub mod tests {
                         }
                         assert_eq!(
                             accum, data_this_pass,
-                            "bins: {bins}, start_bin_this_pass: {start_bin_this_pass}, pass: {pass}, flatten: {flatten_data}, passes: {passes}"
+                            "bins: {}, start_bin_this_pass: {}, pass: {}, flatten: {}, passes: {}",
+                            bins, start_bin_this_pass, pass, flatten_data, passes
                         );
                     }
                 }
@@ -427,9 +392,9 @@ pub mod tests {
         bins: usize,
         start_bin: usize,
     ) {
-        let mut accum: SavedType = (0..bins).map(|_| vec![]).collect();
-        data.drain(..).for_each(|mut x| {
-            x.drain(..).for_each(|item| {
+        let mut accum: SavedType = (0..bins).into_iter().map(|_| vec![]).collect();
+        data.drain(..).into_iter().for_each(|mut x| {
+            x.drain(..).into_iter().for_each(|item| {
                 let bin = bin_calculator.bin_from_pubkey(&item.pubkey);
                 accum[bin - start_bin].push(item);
             })
@@ -446,10 +411,12 @@ pub mod tests {
         let mut ct = 0;
         (
             (0..bins)
+                .into_iter()
                 .map(|bin| {
                     let rnd = rng.gen::<u64>() % (bins as u64);
                     if rnd < count as u64 {
                         (0..std::cmp::max(1, count / bins))
+                            .into_iter()
                             .map(|_| {
                                 ct += 1;
                                 let mut pk;
